pyproject.toml
[tool.poetry]
name = "app"
version = "1.0.0"
description = "Houmer tracking App"
authors = ["Yerko Aguirre"]
[tool.poetry.dependencies]
python = "^3.8"
alembic = "^1.4.1"
argon2-cffi = "^19.2.0"
databases = {version = "^0.2.6", extras = ["postgresql"]}
email_validator = "^1.0.5"
fastapi = {extras = ["email_validator", "requests"], version = "^0.52.0"}
passlib = "^1.7.2"
pyjwt = "^1.7.1"
python-multipart = "^0.0.5"
toolz = "^0.10.0"
uvicorn = "^0.11.3"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
factory_boy = "^2.12.0"
Faker = "^4.0.1"
flake8 = "^3.7.9"
flake8-black = "^0.1.1"
flake8-broken-line = "^0.1.1"
flake8-bugbear = "^20.1.4"
flake8-builtins = "^1.4.2"
flake8-comprehensions = "^3.2.2"
flake8-docstrings = "^1.5.0"
flake8-eradicate = "^0.2.4"
flake8-pytest-style = "^0.2.0"
mccabe = "^0.6.1"
mypy = "^0.761"
pep8-naming = "^0.9.1"
pytest = "^5.3.5"
pytest-asyncio = "^0.10.0"
pytest-cov = "^2.8.1"
pytest-factoryboy = "^2.0.3"
pytest-mock = "^2.0.0"
pytest-sugar = "^0.9.2"
python-dotenv = "^0.12.0"

[tool.poetry.scripts]
seeder = "scripts:seeder"
web_server = "scripts:web_server"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
